% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt2ticks.R
\name{dt2ticks}
\alias{dt2ticks}
\title{Datetime to Timeticks}
\usage{
dt2ticks(dt)
}
\arguments{
\item{dt}{Datetime}
}
\value{
\code{character()}
}
\description{
Transforms Datetime to Timeticks format.
}
\examples{
dt2ticks("2020/07/09 11:00:00")
# "637299000000000000"

}
