% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ticks2dt.R
\name{ticks2dt}
\alias{ticks2dt}
\title{Timeticks to Datetime}
\usage{
ticks2dt(ticks)
}
\arguments{
\item{ticks}{Timeticks}
}
\value{
\code{as.POSIXct()} Datetime
}
\description{
Transforms Timeticks to Datetime format.
}
